Create LTDETR detector
