+++ bv_core/detectors/__init__.py
@@ -2,12 +2,10 @@
 
 from .base_detector import BaseDetector
 from .gazebo_bbox_detector import GazeboBBoxDetector
-from .ml_detector import MLDetector
 
 __all__ = [
     "BaseDetector",
     "GazeboBBoxDetector",
-    "MLDetector",
     "create_detector",
 ]
 
@@ -17,7 +15,7 @@ def create_detector(detector_type: str, **config) -> BaseDetector:
 
     Args:
         detector_type: Type of detector to create. Options:
-            - "ml": ML-based detector using RF-DETR
+            - "ml": ML-based detector using LightlyTrain LTDETR
             - "gazebo_bbox": Gazebo bounding box camera detector
         **config: Configuration parameters passed to the detector constructor.
 
@@ -28,9 +26,13 @@ def create_detector(detector_type: str, **config) -> BaseDetector:
         ValueError: If detector_type is unknown.
     """
     if detector_type == "ml":
+        from .ml_detector import MLDetector
+
         return MLDetector(
-            batch_size=config.get("batch_size", 16),
-            resolution=config.get("resolution", 728),
+            model_path=config.get(
+                "ml_model_path",
+                "/Users/allenthomas/Code/Personal/inference/ltdetr.pt",
+            ),
         )
     elif detector_type == "gazebo_bbox":
         return GazeboBBoxDetector(
