+++ bv_core/detectors/ml_detector.py
@@ -1,200 +1,101 @@
-"""ML-based object detector using RF-DETR."""
+"""ML-based object detector using LightlyTrain LTDETR."""
 
-import os
+from pathlib import Path
 
 import numpy as np
 import rclpy.logging
 import supervision as sv
-import torch
-from PIL import Image
-from rfdetr import RFDETRBase
 
 from .base_detector import BaseDetector
 
 
 class MLDetector(BaseDetector):
-    """Object detector using RF-DETR model with tiled inference."""
+    """Object detector using a LightlyTrain LTDETR checkpoint."""
 
-    def __init__(self, batch_size: int = 16, resolution: int = 728):
+    def __init__(self, model_path: str):
         """Initialize the ML detector.
 
         Args:
-            batch_size: Number of tiles to process in parallel.
-            resolution: Input resolution for the model (tile size).
+            model_path: Filesystem path to the LTDETR checkpoint.
         """
-        self.batch_size = batch_size
-        self.resolution = resolution
+        self.model_path = Path(model_path).expanduser()
         self.model = None
-        self.frame_save_cnt = 0
+        self._logger = rclpy.logging.get_logger("ml_detector")
 
     def start(self) -> None:
-        """Load the RF-DETR model."""
+        """Load the LTDETR model."""
         if self.model is None:
             self.model = self._create_model()
 
     def stop(self) -> None:
         """Release model resources."""
         self.model = None
-        if torch.cuda.is_available():
-            torch.cuda.empty_cache()
 
     def process_frame(
         self,
         frame: np.ndarray,
-        overlap: int = 100,
         threshold: float = 0.5,
         **kwargs,
     ) -> sv.Detections:
-        """Process an image by tiling, batched inference, and global NMS.
+        """Run LTDETR inference on a single frame.
 
         Args:
-            frame: Input image as numpy array (H, W, C).
-            overlap: Overlap between adjacent tiles in pixels.
-            threshold: Detection confidence threshold.
+            frame: Input image as numpy array (H, W, C) in BGR format.
+            threshold: Minimum confidence required to keep a detection.
 
         Returns:
-            Merged detections with NMS applied.
+            Detections surviving confidence filtering.
         """
         if self.model is None:
-            rclpy.logging.get_logger("ml_detector").info("Creating model")
             self.model = self._create_model()
 
-        tiles_with_positions = self._create_tiles(frame, overlap)
-        tiles, offsets = zip(*tiles_with_positions)
+        from PIL import Image
 
-        detections_list = []
-        for i in range(0, len(tiles), self.batch_size):
-            batch = list(tiles[i : i + self.batch_size])
-            dets = self._pad_and_predict(batch, threshold)
-            detections_list.extend(dets)
+        image = Image.fromarray(frame[:, :, ::-1])
+        results = self.model.predict(image)
 
-        all_dets = []
-        for dets, (x_off, y_off) in zip(detections_list, offsets):
-            if len(dets.xyxy) == 0:
-                continue
+        labels = self._to_numpy(results.get("labels"), np.int32)
+        boxes = self._to_numpy(results.get("bboxes"), np.float32).reshape((-1, 4))
+        scores = self._to_numpy(results.get("scores"), np.float32)
 
-            boxes = dets.xyxy.copy()
-            boxes[:, [0, 2]] += x_off
-            boxes[:, [1, 3]] += y_off
-
-            all_dets.append(
-                sv.Detections(
-                    xyxy=boxes,
-                    confidence=dets.confidence,
-                    class_id=dets.class_id,
-                )
-            )
-
-        if not all_dets:
+        if len(labels) == 0 or len(boxes) == 0 or len(scores) == 0:
             return sv.Detections.empty()
 
-        merged = sv.Detections.merge(all_dets)
-        return merged.with_nms(threshold=0.45)
+        keep = scores >= float(threshold)
+        if not np.any(keep):
+            return sv.Detections.empty()
 
-    def annotate_frame(
-        self,
-        frame: np.ndarray,
-        detections: sv.Detections,
-        labels: list,
-    ) -> np.ndarray:
-        """Draw bounding boxes and labels on frame."""
-        annotated_frame = frame.copy()
-        annotated_frame = sv.BoxAnnotator().annotate(annotated_frame, detections)
-        annotated_frame = sv.LabelAnnotator().annotate(
-            annotated_frame, detections, labels
-        )
-        return annotated_frame
-
-    def save_frame(self, frame: np.ndarray, output_dir: str) -> None:
-        """Save annotated frame to disk."""
-        os.makedirs(output_dir, exist_ok=True)
-        output_path = os.path.join(output_dir, f"{self.frame_save_cnt}_annotated.jpg")
-        Image.fromarray(frame).save(output_path)
-        self.frame_save_cnt += 1
-
-    def _create_tiles(self, image: np.ndarray, overlap: int = 100) -> list:
-        """Split an image into overlapping tiles.
-
-        Args:
-            image: Input image as numpy array (H, W, C).
-            overlap: Overlap between adjacent tiles in pixels.
-
-        Returns:
-            List of (tile, (x_offset, y_offset)) tuples.
-        """
-        tile_size = self.resolution
-        width, height = image.shape[1], image.shape[0]
-        tiles = []
-
-        x_positions = list(range(0, width - overlap, tile_size - overlap))
-        if width not in x_positions and x_positions:
-            x_positions.append(max(0, width - tile_size))
-        if not x_positions:
-            x_positions = [0]
-
-        y_positions = list(range(0, height - overlap, tile_size - overlap))
-        if height not in y_positions and y_positions:
-            y_positions.append(max(0, height - tile_size))
-        if not y_positions:
-            y_positions = [0]
-
-        for y in y_positions:
-            for x in x_positions:
-                x_end = min(x + tile_size, width)
-                y_end = min(y + tile_size, height)
-                x_start = max(0, x_end - tile_size)
-                y_start = max(0, y_end - tile_size)
-
-                tile = image[y_start:y_end, x_start:x_end, :]
-                tiles.append((tile, (x_start, y_start)))
-
-        return tiles
-
-    def _pad_and_predict(self, tiles: list, threshold: float) -> list:
-        """Run batched inference, padding if needed.
-
-        Args:
-            tiles: List of tile images as numpy arrays.
-            threshold: Detection confidence threshold.
-
-        Returns:
-            List of detections for each tile.
-        """
-        B = len(tiles)
-        if B < self.batch_size:
-            dummy = np.zeros_like(tiles[0])
-            tiles = tiles + [dummy] * (self.batch_size - B)
-            trim_to = B
-        else:
-            trim_to = None
-
-        outputs = self.model.predict(tiles, threshold=threshold)
-
-        if trim_to is not None:
-            outputs = outputs[:trim_to]
-
-        return outputs
-
-    def _create_model(self, dtype=None):
-        """Initialize the RF-DETR model."""
-        if dtype is None:
-            if torch.cuda.is_available():
-                dtype = torch.float16
-            else:
-                dtype = torch.float32
-
-        if torch.cuda.is_available():
-            torch.cuda.empty_cache()
-            device_info = "cuda"
-        else:
-            device_info = "cpu"
-
-        model = RFDETRBase(resolution=self.resolution)
-        model.optimize_for_inference(batch_size=self.batch_size, dtype=dtype)
-
-        rclpy.logging.get_logger("ml_detector").info(
-            f"Model created with batch size: {self.batch_size} "
-            f"(dtype={dtype}, device={device_info})"
+        return sv.Detections(
+            xyxy=boxes[keep],
+            confidence=scores[keep],
+            class_id=labels[keep],
         )
 
-        return model
+    def _create_model(self):
+        """Initialize the LTDETR model."""
+        if not self.model_path.is_file():
+            raise RuntimeError(f"LTDETR weights not found: {self.model_path}")
+
+        try:
+            import lightly_train
+        except ImportError as exc:
+            raise RuntimeError(
+                "Missing dependency 'lightly_train'. Install it with "
+                "`python3 -m pip install lightly-train`."
+            ) from exc
+
+        self._logger.info(f"Loading LTDETR model from {self.model_path}")
+        return lightly_train.load_model(str(self.model_path))
+
+    @staticmethod
+    def _to_numpy(value, dtype) -> np.ndarray:
+        """Convert LTDETR outputs to numpy arrays without hard-coding torch."""
+        if value is None:
+            return np.array([], dtype=dtype)
+        if hasattr(value, "detach"):
+            value = value.detach()
+        if hasattr(value, "cpu"):
+            value = value.cpu()
+        if hasattr(value, "numpy"):
+            value = value.numpy()
+        return np.asarray(value, dtype=dtype)
