+++ config/vision_params.yaml
@@ -3,16 +3,14 @@
 # ============================================================
 
 # Detection settings
-batch_size: 4
 detection_threshold: 0.5
-resolution: 560
-overlap: 100
 capture_interval: 1.5e9
 estimated_camera_latency: 0.5  # Seconds to pause after stopping before capturing frame
 num_scan_wp: 3
 
 # Detector configuration
-detector_type: "gazebo_bbox"  # Options: "ml" (RF-DETR), "gazebo_bbox" (Gazebo bounding box camera)
+detector_type: "gazebo_bbox"  # Options: "ml" (LightlyTrain LTDETR), "gazebo_bbox" (Gazebo bounding box camera)
+ml_model_path: "ltdetr.pt"  # LTDETR checkpoint path when detector_type is "ml"
 gazebo_bbox_topic: "/camera/bounding_boxes"  # Topic for Gazebo bbox camera (only used when detector_type is "gazebo_bbox")
 
 # ============================================================
