+++ requirements.txt
@@ -1,4 +1,6 @@
 supervision
+lightly-train
+Pillow
 empy==3.3.4
 pyros-genmsg 
 setuptools==65.5.1
